Added check for older msfvenom payloads
The older msfvenom payload gets flagged by AVG (1/35 on nodistribute). Began looking for the signatures and found it being flagged off of the string WWWW in the original Payload.smali and PayloadTrustManager.smali. A sed command could fix this, but AVG also has a signature based on the bytesToHex method in PayloadTrustManager that became too much of a pain to bother with. Easy fix is to force user's to utilize the new version as well as it being hard to support multiple versions in the future as signatures expand.