Different placement for permission
Got a few new AV signatures today based on the placement of the permission so simply moving it to another spot avoided it.

Future update will place all permissions into an array and add them in randomly. If they start flagging based on the permissions seen, then may give options for restricted permissions to avoid AV flags. Or the ability to set specific permissions on/off.