Resolved of option '-o' bug + APKTool latest
Resolved a bug regarding the -o option. The issue was related to not overwriting default filenames for other variables that were used. 

Modified the auto-download of APKTool to use 2.2.4 vs the older 2.2.2 for the latest improvements. This only applies to those whom don't already have APKTool configured. May look into checking the version and auto-updating in the future.